+++ src/entity/CoworkingSpace.java
@@ -1,3 +1,7 @@
+package entity;
+
+import enums.TypeOfWorkspaces;
+
 public class CoworkingSpace {
     private static Integer nextId = 1;
 
@@ -18,7 +22,7 @@ public class CoworkingSpace {
         this.id = nextId++;
     }
 
-    public CoworkingSpace(){
+    public CoworkingSpace() {
         this.id = nextId++;
     }
 
@@ -60,9 +64,9 @@ public class CoworkingSpace {
 
     @Override
     public String toString() {
-        return  "(id)=" + id +
+        return "(id)=" + id +
                 ", " + name +
-                 ", typeOfWorkspaces=" + typeOfWorkspaces +
+                ", typeOfWorkspaces=" + typeOfWorkspaces +
                 ", price=" + priceInDollars +
                 ", availabilityStatus=" + availabilityStatus;
     }
