+++ src/entity/menus/Menu.java
@@ -1,3 +1,5 @@
+package entity.menus;
+
 public abstract class Menu {
     public abstract void showMenu();
 }
