+++ src/entity/roles/Admin.java
@@ -1,17 +1,23 @@
+package entity.roles;
+
+import entity.CoworkingSpace;
+import utils.BookingManagement;
+import utils.WorkspaceManagement;
+
 public class Admin extends User {
 
     public void addCoworkingSpace(CoworkingSpace coworkingSpace) {
         WorkspaceManagement.allCoworkingSpaces.add(coworkingSpace);
     }
 
-    public boolean removeCoworkingSpace(int id){
-       CoworkingSpace coworkingSpace = WorkspaceManagement.getCoworkingSpaceById(id);
-       if(coworkingSpace == null) {
-           return false;
-       }else {
-           WorkspaceManagement.allCoworkingSpaces.remove(coworkingSpace);
-           return true;
-       }
+    public boolean removeCoworkingSpace(int id) {
+        CoworkingSpace coworkingSpace = WorkspaceManagement.getCoworkingSpaceById(id);
+        if (coworkingSpace == null) {
+            return false;
+        } else {
+            WorkspaceManagement.allCoworkingSpaces.remove(coworkingSpace);
+            return true;
+        }
     }
