+++ src/entity/roles/Customer.java
@@ -1,3 +1,9 @@
+package entity.roles;
+
+import entity.Booking;
+import utils.BookingManagement;
+import utils.WorkspaceManagement;
+
 import java.util.HashSet;
 
 public class Customer extends User {
@@ -8,24 +14,24 @@ public class Customer extends User {
 
     public boolean cancelReservation(int id) {
         Booking booking = BookingManagement.getBookingById(id);
-        if(booking == null) {
+        if (booking == null) {
             return false;
-        }else {
+        } else {
             BookingManagement.allBookings.remove(booking);
             customerReservations.remove(booking);
             return true;
         }
     }
 
-    public void makeReservation(Booking booking){
+    public void makeReservation(Booking booking) {
         customerReservations.add(booking);
         BookingManagement.allBookings.add(booking);
     }
 
     public void viewReservations() {
-        for (Booking booking: customerReservations) {
-            System.out.println(booking.getId() + "\n" + booking.getCustomerName() + "\n" + booking.getDate() + " " + booking.getStartAndEndOfBookingTime() +"\n"
-            + booking.getIdOfCoworkingSpace() + " " + WorkspaceManagement.getCoworkingSpaceById(booking.getIdOfCoworkingSpace()).getName());
+        for (Booking booking : customerReservations) {
+            System.out.println(booking.getId() + "\n" + booking.getCustomerName() + "\n" + booking.getDate() + " " + booking.getStartAndEndOfBookingTime() + "\n"
+                    + booking.getIdOfCoworkingSpace() + " " + WorkspaceManagement.getCoworkingSpaceById(booking.getIdOfCoworkingSpace()).getName());
             System.out.println("------------------------------");
         }
     }
