+++ src/entity/roles/User.java
@@ -1,3 +1,5 @@
+package entity.roles;
+
 public class User {
     private String login;
